requirement.txt
llama-cpp-python==0.2.85
fastapi
uvicorn
huggingface_hub
aiohttp